chore(trace-viewer): Reorganise files
