+++ src/utils/browserFetcher.ts
@@ -19,8 +19,7 @@ import extract from 'extract-zip';
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
-import { existsAsync } from './utils';
-import { download } from './download';
+import { existsAsync, download } from './utils';
 import { debugLogger } from './debugLogger';
 
 export async function downloadBrowserWithProgressBar(title: string, browserDirectory: string, executablePath: string, downloadURL: string, downloadFileName: string): Promise<boolean> {
@@ -36,6 +35,7 @@ export async function downloadBrowserWithProgressBar(title: string, browserDirec
   try {
     await download(url, zipPath, {
       progressBarName,
+      log: debugLogger.log.bind(debugLogger, 'install')
     });
     debugLogger.log('install', `extracting archive`);
     debugLogger.log('install', `-- zip: ${zipPath}`);
