--- src/utils/download.ts
@@ -1,144 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import ProgressBar from 'progress';
-import { debugLogger } from './debugLogger';
-import { getErrorMessage, httpRequest } from './utils';
-import * as fs from 'fs';
-
-
-type OnProgressCallback = (downloadedBytes: number, totalBytes: number) => void;
-type DownloadFileLogger = (message: string) => void;
-
-function toMegabytes(bytes: number): string {
-  const mb = bytes / 1024 / 1024;
-  return `${Math.round(mb * 10) / 10} Mb`;
-}
-
-function getDownloadProgress(progressBarName: string): OnProgressCallback {
-  let progressBar: ProgressBar;
-  let lastDownloadedBytes = 0;
-
-  return (downloadedBytes: number, totalBytes: number) => {
-    if (!process.stderr.isTTY) return;
-    if (!progressBar) {
-      progressBar = new ProgressBar(
-          `Downloading ${progressBarName} - ${toMegabytes(
-              totalBytes
-          )} [:bar] :percent :etas `,
-          {
-            complete: '=',
-            incomplete: ' ',
-            width: 20,
-            total: totalBytes,
-          }
-      );
-    }
-    const delta = downloadedBytes - lastDownloadedBytes;
-    lastDownloadedBytes = downloadedBytes;
-    progressBar.tick(delta);
-  };
-}
-
-function downloadFile(
-  url: string,
-  destinationPath: string,
-  options: {
-    progressCallback?: OnProgressCallback;
-    log?: DownloadFileLogger;
-  } = {}
-): Promise<{ error: any }> {
-  const { progressCallback, log = () => {} } = options;
-  log(`running download:`);
-  log(`-- from url: ${url}`);
-  log(`-- to location: ${destinationPath}`);
-  let fulfill: ({ error }: { error: any }) => void = ({ error }) => {};
-  let downloadedBytes = 0;
-  let totalBytes = 0;
-
-  const promise: Promise<{ error: any }> = new Promise(x => {
-    fulfill = x;
-  });
-
-  httpRequest(
-      { url },
-      response => {
-        log(`-- response status code: ${response.statusCode}`);
-        if (response.statusCode !== 200) {
-          const error = new Error(
-              `Download failed: server returned code ${response.statusCode}. URL: ${url}`
-          );
-          // consume response data to free up memory
-          response.resume();
-          fulfill({ error });
-          return;
-        }
-        const file = fs.createWriteStream(destinationPath);
-        file.on('finish', () => fulfill({ error: null }));
-        file.on('error', error => fulfill({ error }));
-        response.pipe(file);
-        totalBytes = parseInt(response.headers['content-length'] || '0', 10);
-        log(`-- total bytes: ${totalBytes}`);
-        if (progressCallback) response.on('data', onData);
-      },
-      (error: any) => fulfill({ error })
-  );
-  return promise;
-
-  function onData(chunk: string) {
-    downloadedBytes += chunk.length;
-    progressCallback!(downloadedBytes, totalBytes);
-  }
-}
-
-export async function download(
-  url: string,
-  destination: string,
-  options: {
-		progressBarName?: string,
-		retryCount?: number
-	} = {}
-) {
-  const { progressBarName = 'file', retryCount = 3 } = options;
-  for (let attempt = 1; attempt <= retryCount; ++attempt) {
-    debugLogger.log(
-        'download',
-        `downloading ${progressBarName} - attempt #${attempt}`
-    );
-    const { error } = await downloadFile(url, destination, {
-      progressCallback: getDownloadProgress(progressBarName),
-      log: debugLogger.log.bind(debugLogger, 'download'),
-    });
-    if (!error) {
-      debugLogger.log('download', `SUCCESS downloading ${progressBarName}`);
-      break;
-    }
-    const errorMessage = getErrorMessage(error);
-    debugLogger.log('download', `attempt #${attempt} - ERROR: ${errorMessage}`);
-    if (
-      attempt < retryCount &&
-      (errorMessage.includes('ECONNRESET') ||
-        errorMessage.includes('ETIMEDOUT'))
-    ) {
-      // Maximum default delay is 3rd retry: 1337.5ms
-      const millis = Math.random() * 200 + 250 * Math.pow(1.5, attempt);
-      debugLogger.log('download', `sleeping ${millis}ms before retry...`);
-      await new Promise(c => setTimeout(c, millis));
-    } else {
-      throw error;
-    }
-  }
-}
